% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{cong_bill}
\alias{cong_bill}
\title{Request Bill Information}
\usage{
cong_bill(
  congress = NULL,
  type = NULL,
  number = NULL,
  item = NULL,
  format = "json",
  clean = TRUE
)
}
\arguments{
\item{congress}{Congress number to search for. 81 or later are supported.}

\item{type}{Type of bill. Can be \code{'hr'}, \code{'s'}, \code{'hjres'}, \code{'sjres'}, \code{'hconres'}, \code{'sconres'}, \code{'hres'}, or \code{'sres'}.}

\item{number}{Bill assigned number. Numeric.}

\item{item}{Information to request. Can be \code{'actions'}, \code{'amendments'}, \code{'committees'}, \code{'cosponsors'},
\code{'relatedbills'}, \code{'subjects'}, \code{'text'}, \code{'titles'}}

\item{format}{Output format for \code{clean = FALSE}. One of \code{xml} or \code{json}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or requested \code{format}s.}
}
\value{
\code{tibble} or HTTP response.
}
\description{
Request Bill Information
}
\examples{
\dontrun{
# Requires API Key

cong_bill()

cong_bill(congress = 117)

cong_bill(congress = 117, type = 'hr', number = 3076, clean = F)


}

}
