% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_next.R
\name{cong_request_next}
\alias{cong_request_next}
\title{Request next set of responses}
\usage{
cong_request_next(response, max_req = 1)
}
\arguments{
\item{response}{A \link[tibble:tibble]{tibble::tibble} from a \verb{cong_*} function}

\item{max_req}{A max number of additional requests to make. Default is 1.}
}
\value{
a tibble with responses bound by row to new results
}
\description{
Request next set of responses
}
\examples{
\dontshow{if (congress::has_congress_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API Key

cong_bill() |>
  cong_request_next()
\dontshow{\}) # examplesIf}
}
