% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treaty.R
\name{cong_treaty}
\alias{cong_treaty}
\title{Request Treaty Information}
\usage{
cong_treaty(
  congress = NULL,
  number = NULL,
  suffix = NULL,
  item = NULL,
  from_date = NULL,
  to_date = NULL,
  limit = 20,
  offset = 0,
  format = "json",
  clean = TRUE
)
}
\arguments{
\item{congress}{Congress number to search for. 81 or later are supported.}

\item{number}{Treaty assigned number. Numeric.}

\item{suffix}{Treaty partition letter value. Character.}

\item{item}{Information to request. Can be \code{'actions'}, or \code{'committees'}.}

\item{from_date}{start date for search, e.g. \code{'2022-04-01'}. Defaults to most recent.}

\item{to_date}{end date for search, e.g. \code{'2022-04-03'}. Defaults to most recent.}

\item{limit}{number of records to return. Default is 20. Will be truncated to between 1 and 250.}

\item{offset}{number of records to skip. Default is 0. Must be non-negative.}

\item{format}{Output format for \code{clean = FALSE}. One of \code{xml} or \code{json}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or requested \code{format}.}
}
\value{
\code{tibble} or HTTP response if \code{clean = FALSE}
}
\description{
Request Treaty Information
}
\examples{
\dontshow{if (congress::has_congress_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Requires API Key

cong_treaty()

cong_treaty(congress = 117)

cong_treaty(congress = 117, number = 3)

cong_treaty(congress = 114, number = 13, suffix = 'A')

cong_treaty(congress = 117, number = 3, item = 'actions')

cong_treaty(congress = 114, number = 13, suffix = 'A', item = 'actions')
\dontshow{\}) # examplesIf}
}
