% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congress}
\alias{congress}
\title{US Congress}
\format{A tibble with 21,009 rows and 32 columns.

\describe{
\item{\code{congress}}{Congressional session.}
\item{\code{pid}}{Unique Person ID.}
\item{\code{last}}{Last name.}
\item{\code{first}}{First name.}
\item{\code{middle}}{Middle name.}
\item{\code{suffix}}{Name suffix.}
\item{\code{born}}{Date of birth}
\item{\code{death}}{Date of death.}
\item{\code{sex}}{Sex.}
\item{\code{position}}{Position.}
\item{\code{party}}{Party affiliation.}
\item{\code{state}}{State represented.}
\item{\code{district}}{District number.}
\item{\code{start}}{Year first elected to role.}
\item{\code{end}}{Final year served in role.}
\item{\code{religion}}{Religion.}
\item{\code{race}}{Race.}
\item{\code{education}}{Educational attainment.}
\item{\code{occ1}}{Most recent occupation type.}
\item{\code{occ2}}{Second most recent occupation type.}
\item{\code{occ3}}{Third most recent occupation type.}
\item{\code{mil1}}{Most recent military service type.}
\item{\code{mil2}}{Second most recent military service type.}
\item{\code{mil3}}{Third most recent military service type.}
\item{\code{start_year}}{Start of present term (i.e. term described by this row).}
\item{\code{end_year}}{End of present term (i.e. term described by this row).}
\item{\code{start_age}}{Age when beginning present term (i.e. term described by this row).}
\item{\code{poc}}{Person of color.}
\item{\code{days_old}}{Days of age when beginning present term (i.e. term described by this row).}
\item{\code{entry_age}}{Age when first elected to position, in years.}
\item{\code{end_career}}{End date of final term.}
\item{\code{yr_fac}}{Same as start_year but coded as a factor instead of a date.}
}}
\source{
.
}
\usage{
congress
}
\description{
US Congressional Representatives since 1945 (House, Senate, Presidency, Vice-Presidency)
}
\keyword{datasets}
