% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_key.R
\name{key}
\alias{key}
\alias{has_congress_key}
\alias{get_congress_key}
\title{Check or Get Congress API Key}
\usage{
has_congress_key()

get_congress_key()
}
\value{
logical if \code{has}, key if \code{get}
}
\description{
Check or Get Congress API Key
}
\examples{
has_congress_key()
}
