% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_key.R
\name{set_congress_key}
\alias{set_congress_key}
\title{Add Entry to Renviron}
\usage{
set_congress_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{Character. API key to add to add.}

\item{overwrite}{Defaults to FALSE. Boolean. Should existing \code{CONGRESS_KEY} in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Boolean. Should this be added '~/.Renviron' file?}
}
\value{
key, invisibly
}
\description{
Adds Congress API key to .Renviron.
}
\examples{
\dontrun{
set_congress_key('1234')
}

}
