% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member.R
\name{cong_member}
\alias{cong_member}
\title{Request Member Information}
\usage{
cong_member(
  bioguide = NULL,
  item = NULL,
  from_date = NULL,
  to_date = NULL,
  limit = 20,
  offset = 0,
  format = "json",
  clean = TRUE
)
}
\arguments{
\item{bioguide}{Bioguide identifier for a member of Congress.}

\item{item}{Information to request. Can be \code{'sponsored-legislation'} or  \code{'cosponsored-legislation'}}

\item{from_date}{start date for search, e.g. \code{'2022-04-01'}. Defaults to most recent.}

\item{to_date}{end date for search, e.g. \code{'2022-04-03'}. Defaults to most recent.}

\item{limit}{number of records to return. Default is 20. Will be truncated to between 1 and 250.}

\item{offset}{number of records to skip. Default is 0. Must be non-negative.}

\item{format}{Output format for \code{clean = FALSE}. One of \code{xml} or \code{json}.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or requested \code{format}.}
}
\value{
\code{tibble} or HTTP response if \code{clean = FALSE}
}
\description{
Request Member Information
}
\examples{
\dontrun{
# Requires API Key

cong_member()

cong_member(bioguide = 'L000174', clean = FALSE)

cong_member(bioguide = 'L000174', item = 'sponsored-legislation')

}

}
